# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Raízes do Nordeste API",
    "description": "API acadêmica para a rede Raízes do Nordeste, com multicanalidade, estoque por unidade, pedido, pagamento mock, fidelização, segurança e auditoria.",
    "contact": {
      "name": "Projeto Multidisciplinar - Trilha Back-End"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Sistema"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/auth/cadastro": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Cadastro",
        "operationId": "cadastro_auth_cadastro_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCadastro"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenSaida"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/unidades": {
      "get": {
        "tags": [
          "Catálogo"
        ],
        "summary": "Listar Unidades",
        "operationId": "listar_unidades_unidades_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/produtos": {
      "get": {
        "tags": [
          "Catálogo"
        ],
        "summary": "Listar Produtos",
        "operationId": "listar_produtos_produtos_get",
        "parameters": [
          {
            "name": "unidadeId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Unidade cujo cardápio/estoque será consultado",
              "title": "Unidadeid"
            },
            "description": "Unidade cujo cardápio/estoque será consultado"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/promocoes": {
      "get": {
        "tags": [
          "Catálogo"
        ],
        "summary": "Listar Promocoes",
        "operationId": "listar_promocoes_promocoes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/pedidos": {
      "post": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Novo Pedido",
        "operationId": "novo_pedido_pedidos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PedidoCriacao"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Listar Pedidos",
        "operationId": "listar_pedidos_pedidos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "canalPedido",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CanalPedido"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Canalpedido"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StatusPedido"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pedidos/{pedido_id}": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Obter Pedido",
        "operationId": "obter_pedido_pedidos__pedido_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pedido_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pedido Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pedidos/{pedido_id}/status": {
      "patch": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Atualizar Status",
        "operationId": "atualizar_status_pedidos__pedido_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pedido_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pedido Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusPedidoEntrada"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pagamentos/{pedido_id}/processar": {
      "post": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Processar",
        "operationId": "processar_pagamentos__pedido_id__processar_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pedido_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pedido Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PagamentoEntrada"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/estoque/unidades/{unidade_id}": {
      "get": {
        "tags": [
          "Estoque"
        ],
        "summary": "Saldo",
        "operationId": "saldo_estoque_unidades__unidade_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unidade_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Unidade Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/estoque/movimentacoes": {
      "post": {
        "tags": [
          "Estoque"
        ],
        "summary": "Movimentar",
        "operationId": "movimentar_estoque_movimentacoes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovimentoEstoqueEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/fidelidade/saldo": {
      "get": {
        "tags": [
          "Fidelidade e Auditoria"
        ],
        "summary": "Saldo Fidelidade",
        "operationId": "saldo_fidelidade_fidelidade_saldo_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auditoria": {
      "get": {
        "tags": [
          "Fidelidade e Auditoria"
        ],
        "summary": "Auditoria",
        "operationId": "auditoria_auditoria_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "CanalPedido": {
        "type": "string",
        "enum": [
          "APP",
          "TOTEM",
          "BALCAO",
          "PICKUP",
          "WEB"
        ],
        "title": "CanalPedido"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ItemPedidoEntrada": {
        "properties": {
          "produtoId": {
            "type": "integer",
            "title": "Produtoid"
          },
          "quantidade": {
            "type": "integer",
            "maximum": 50.0,
            "exclusiveMinimum": 0.0,
            "title": "Quantidade"
          }
        },
        "type": "object",
        "required": [
          "produtoId",
          "quantidade"
        ],
        "title": "ItemPedidoEntrada"
      },
      "LoginEntrada": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "senha": {
            "type": "string",
            "title": "Senha"
          }
        },
        "type": "object",
        "required": [
          "email",
          "senha"
        ],
        "title": "LoginEntrada"
      },
      "MovimentoEstoqueEntrada": {
        "properties": {
          "unidadeId": {
            "type": "integer",
            "title": "Unidadeid"
          },
          "produtoId": {
            "type": "integer",
            "title": "Produtoid"
          },
          "quantidade": {
            "type": "integer",
            "title": "Quantidade"
          },
          "motivo": {
            "type": "string",
            "maxLength": 120,
            "minLength": 3,
            "title": "Motivo"
          }
        },
        "type": "object",
        "required": [
          "unidadeId",
          "produtoId",
          "quantidade",
          "motivo"
        ],
        "title": "MovimentoEstoqueEntrada"
      },
      "PagamentoEntrada": {
        "properties": {
          "resultado": {
            "$ref": "#/components/schemas/StatusPagamento"
          }
        },
        "type": "object",
        "required": [
          "resultado"
        ],
        "title": "PagamentoEntrada"
      },
      "PedidoCriacao": {
        "properties": {
          "unidadeId": {
            "type": "integer",
            "title": "Unidadeid"
          },
          "canalPedido": {
            "$ref": "#/components/schemas/CanalPedido"
          },
          "itens": {
            "items": {
              "$ref": "#/components/schemas/ItemPedidoEntrada"
            },
            "type": "array",
            "minItems": 1,
            "title": "Itens"
          },
          "formaPagamento": {
            "type": "string",
            "title": "Formapagamento",
            "default": "MOCK"
          }
        },
        "type": "object",
        "required": [
          "unidadeId",
          "canalPedido",
          "itens"
        ],
        "title": "PedidoCriacao"
      },
      "StatusPagamento": {
        "type": "string",
        "enum": [
          "PENDENTE",
          "APROVADO",
          "RECUSADO"
        ],
        "title": "StatusPagamento"
      },
      "StatusPedido": {
        "type": "string",
        "enum": [
          "AGUARDANDO_PAGAMENTO",
          "PAGO",
          "EM_PREPARO",
          "PRONTO",
          "ENTREGUE",
          "CANCELADO",
          "PAGAMENTO_RECUSADO"
        ],
        "title": "StatusPedido"
      },
      "StatusPedidoEntrada": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/StatusPedido"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StatusPedidoEntrada"
      },
      "TokenSaida": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken"
          },
          "tokenType": {
            "type": "string",
            "title": "Tokentype",
            "default": "Bearer"
          },
          "expiresIn": {
            "type": "integer",
            "title": "Expiresin"
          },
          "user": {
            "additionalProperties": true,
            "type": "object",
            "title": "User"
          }
        },
        "type": "object",
        "required": [
          "accessToken",
          "expiresIn",
          "user"
        ],
        "title": "TokenSaida"
      },
      "UsuarioCadastro": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Nome"
          },
          "email": {
            "type": "string",
            "maxLength": 180,
            "minLength": 5,
            "title": "Email"
          },
          "senha": {
            "type": "string",
            "maxLength": 72,
            "minLength": 8,
            "title": "Senha"
          },
          "consentimentoFidelidade": {
            "type": "boolean",
            "title": "Consentimentofidelidade",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "nome",
          "email",
          "senha"
        ],
        "title": "UsuarioCadastro"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}